requirements.txt
kafka-python==2.0.2
psycopg2==2.9.1
SQLAlchemy==1.3.24
